+++ sphinx/source/changelog.rst
@@ -5,6 +5,12 @@ Changelog
 ------
 * Dropped support for Python 3.9
 * Added support for the MEMPTR register to CMIOSimulator and CCMIOSimulator
+* :ref:`snapinfo.py` now shows the MEMPTR register value in SZX snapshots
+* Added support to :ref:`trace.py` for setting the MEMPTR register before
+  execution begins, and for reading and writing the MEMPTR register in SZX
+  snapshots
+* Added support to :ref:`bin2sna.py`, :ref:`snapmod.py` and :ref:`tap2sna.py`
+  for setting the MEMPTR register in SZX snapshots
 * Fixed the bug in CSimulator that restricts the ``timeout`` simulated load
   configuration parameter to a maximum value of 1227 seconds